[FIX] No such field(s) in model mgmtsystem.nonconformity
get default risk_computation_id from object ref

use proper function args for get_object_reference

[FIX] ValueError: No such external ID currently defined in the system: mgmtsystem_hazard.risk_computation_a_times_b_times_c

[ADD] Split models

[FIX] PEP8 error + cleanup

[FIX] Method name

[FIX] ValueError: No such external ID currently defined in the system: mgmtsystem_hazard.risk_computation_a_times_b_times_c
