+++ mgmtsystem_hazard/models/__init__.py
@@ -1,0 +1,6 @@
+# -*- encoding: utf-8 -*-
+
+from . import (
+    mgmtsystem_hazard,
+    res_company
+)
