+++ mgmtsystem_hazard/models/mgmtsystem_hazard.py
@@ -48,27 +48,6 @@ class mgmtsystem_hazard_risk_computation(orm.Model):
     }
 
 
-class res_company(orm.Model):
-    _inherit = "res.company"
-    _columns = {
-        'risk_computation_id': fields.many2one(
-            'mgmtsystem.hazard.risk.computation',
-            'Risk Computation',
-            required=True,
-        ),
-    }
-
-    def _get_formula(self, cr, uid, context=None):
-        ids = self.pool.get('mgmtsystem.hazard.risk.computation').search(
-            cr, uid, [('name', '=', 'A * B * C')], context=context
-        )
-        return ids and ids[0] or False
-
-    _defaults = {
-        'risk_computation_id': _get_formula
-    }
-
-
 class mgmtsystem_hazard_risk_type(orm.Model):
 
     _name = "mgmtsystem.hazard.risk.type"
